write multi-column with arrow-json
